Stock decimal

+++ BusinessLogic/DTOs/DTOsProductItem/ProductItemResponse.cs
@@ -5,8 +5,8 @@ namespace BusinessLogic.DTOs.DTOsProductItem
 {
     public class ProductItemResponse
     {
-        public int Quantity { get; set; }
-        public int? Weight { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal? Weight { get; set; }
         public decimal UnitPrice { get; set; }
         public decimal Discount { get; set; }
         public ProductResponse? Product { get; set; }
